kafka/publisher/requirements.txt
python-dotenv==1.0.1
kafka-python==2.2.7
psutil==7.0.0